Update README.md
Snake is the common name for a video game concept where the player maneuvers a line which grows in length, with the line itself being a primary obstacle.

In the game of Snake, the player uses the arrow keys to move a "snake" around the board. As the snake finds food, it eats the food, and thereby grows larger. The game ends when the snake either moves off the screen or moves into itself. The goal is to make the snake as large as possible before that happens